+++ .gitignore
@@ -1,4 +1,7 @@
 .DS_Store
 .ipynb_checkpoints
 *.rdb
-env/
+venv/
+*.egg-info/
+test/
+__pycache__/
