+++ req.txt
@@ -1,0 +1,19 @@
+certifi==2022.12.7
+charset-normalizer==3.1.0
+click==8.1.3
+dash==2.9.1
+dash-bootstrap-components==1.4.1
+dash-core-components==2.0.0
+dash-html-components==2.0.0
+dash-table==5.0.0
+Flask==2.2.3
+idna==3.4
+itsdangerous==2.1.2
+Jinja2==3.1.2
+MarkupSafe==2.1.2
+plotly==5.13.1
+redis==4.5.3
+requests==2.28.2
+tenacity==8.2.2
+urllib3==1.26.15
+Werkzeug==2.2.3
